requirements.txt
dnspython==2.7.0
pymongo==4.11.1
certifi==2025.1.31
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
PyYAML==6.0.2
dill==0.3.9
imbalanced-learn==0.13.0
xgboost==2.1.4
fastapi==0.115.8
httptools==0.6.4
mypy-boto3-s3==1.36.21
types-s3transfer==0.11.2
uvicorn==0.34.0
starlette==0.45.3

-e .
